+++ custom_components/glance_clock/__init__.py
@@ -16,6 +16,8 @@ PLATFORMS = [
     Platform.SWITCH,
     Platform.LIGHT,
     Platform.SELECT,
+    Platform.NUMBER,
+    Platform.BUTTON,
 ]